+++ headers.go
@@ -43,6 +43,10 @@ const (
 	HEADER_OWNER        Header = `"owner"`
 	HEADER_OWNER_UPDATE Header = `"ownerCanUpdate"`
 
+	HEADER_NAME_V2        Header = `"var_header_name"`
+	HEADER_DESCRIPTION_V2 Header = `"var_header_description"`
+	HEADER_ICON_URL_V2    Header = `"var_header_icon"`
+
 	LINE_MODS_STORE    = `34 STORE("mods", `
 	LINE_DOC_VERSION   = `50 STORE("docVersion", `
 	LINE_INDEX_VERSION = `50 STORE("telaVersion", `